requirements.txt
Flask>=2.2.0
spotipy>=2.23.0
python-dotenv>=1.0.0
